+++ src/react-apps/applications/service-development/src/sagas/index.ts
@@ -5,7 +5,7 @@ import { sagaMiddleware } from '../store';
 import { watchHandleFetchInitialCommitSaga, watchHandleFetchServiceConfigSaga, watchHandleFetchServiceNameSaga, watchHandleFetchServiceSaga, watchHandleSaveServiceConfigSaga, watchHandleSaveServiceNameSaga } from '../features/administration/handleServiceInformationSagas';
 import { deploySagas } from '../features/deploy/deploySagas';
 import { watchHandleMergeConflictSaga } from '../features/handleMergeConflict/handleMergeConflictSagas';
-import AppDeploymentSagas from '../sharedResources/appDeploy/appDeploySagas';
+import { appDeploymentSagas } from '../sharedResources/appDeploy/appDeploySagas';
 import { applicationMetadataSagas } from '../sharedResources/applicationMetadata/applicationMetadataSagas';
 import { appReleaseSagas } from '../sharedResources/appRelease/appReleaseSagas';
 import languageSagas from '../utils/fetchLanguage/languageSagas';
@@ -26,7 +26,7 @@ function* root(): SagaIterator {
   yield fork(appClusterSagas);
   yield fork(repoStatusSagas);
   yield fork(appReleaseSagas);
-  yield fork(AppDeploymentSagas);
+  yield fork(appDeploymentSagas);
 }
 
 export const run: () => Task = () => sagaMiddleware.run(root);
